requirements.txt
amqp==2.1.4
appdirs==1.4.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.2
celery==4.0.2
decorator==4.0.11
Django==1.10.6
django-redis==4.7.0
djangorestframework==3.6.1
enum34==1.1.6
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.1.0
kombu==4.0.2
olefile==0.44
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
psycopg2==2.7
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
pytz==2016.10
redis==2.10.5
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
vine==1.1.3
wcwidth==0.1.7
